MDLSITE-2047 Add --branch support to git package
With this new optional option the git package becomes
more deterministic and, in situations where it was
calculating the "best" branch to decide about a given
package, now the passed branch is used instead.

Also, it makes the process more robust because the
commit is verified to be contained in the branch, aborting if not.